linked the database

+++ MvcHotelReservation/Models/Chambre.cs
@@ -9,7 +9,7 @@ public class Chambre
     public string NumeroChambre { get; set; }
     public string TypeChambre { get; set; } // "simple", "double", "suite", "familiale", "standard", "deluxe"
     public int? Capacite { get; set; }
-    public decimal? PrixParNuit { get; set; }
+    public int? PrixParNuit { get; set; }
     public string Description { get; set; }
     public bool Disponibilite { get; set; } = true;
 }
